Data.ASN1.Types

Description

Raw types

data ASN1Class

Element class

Constructors

Universal

Application

Context

Private

type ASN1Tag = Int

ASN1 Tag

data ASN1Length

ASN1 Length with all different formats

Constructors

LenShort Int
Short form with only one byte. length has to be < 127.

LenLong Int Int
Long form of N bytes

LenIndefinite
Length is indefinite expect an EOC in the stream to finish the type

data ASN1Header

ASN1 Header with the class, tag, constructed flag and length.

Constructors

ASN1Header !ASN1Class !ASN1Tag !Bool !ASN1Length

Errors types

data ASN1Error

Possible errors during parsing operations

Constructors

StreamUnexpectedEOC
Unexpected EOC in the stream.

StreamInfinitePrimitive
Invalid primitive with infinite length in a stream.

StreamConstructionWrongSize
A construction goes over the size specified in the header.

StreamUnexpectedSituation String
An unexpected situation has come up parsing an ASN1 event stream.

ParsingHeaderFail String
Parsing an invalid header.

ParsingPartial
Parsing is not finished, there is construction unended.

TypeNotImplemented String
Decoding of a type that is not implemented. Contribution welcome.

TypeDecodingFailed String
Decoding of a knowed type failed.

PolicyFailed String String
Policy failed including the name of the policy and the reason.

Events types

data ASN1Event

represent one event from an asn1 data stream

Constructors

Header ASN1Header
ASN1 Header

Primitive !ByteString
Primitive

ConstructionBegin
Constructed value start

ConstructionEnd
Constructed value end